move conv2d operations to tf package
