+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/Conv2DBackpropFilter.scala
@@ -20,7 +20,7 @@ import java.nio.ByteOrder
 import com.intel.analytics.bigdl.Module
 import com.intel.analytics.bigdl.nn.Identity
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity, DataFormat}
-import com.intel.analytics.bigdl.nn.ops.Conv2DBackFilter
+import com.intel.analytics.bigdl.nn.tf.Conv2DBackFilter
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
