+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/DepthwiseConv2dNative.scala
@@ -19,7 +19,7 @@ import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity, DataFormat}
-import com.intel.analytics.bigdl.nn.ops.{Conv2D, DepthwiseConv2D}
+import com.intel.analytics.bigdl.nn.ops.DepthwiseConv2D
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
 import org.tensorflow.framework.NodeDef
